+++ matrix/NUMA_dense_matrix.h
@@ -141,10 +141,10 @@ public:
 		return data.size();
 	}
 
-	const char *get_row(off_t row_idx) const;
-	char *get_row(off_t row_idx);
-	const char *get_rows(off_t row_start, off_t row_end) const;
-	char *get_rows(off_t row_start, off_t row_end);
+	const char *get_row(size_t row_idx) const;
+	char *get_row(size_t row_idx);
+	const char *get_rows(size_t row_start, size_t row_end) const;
+	char *get_rows(size_t row_start, size_t row_end);
 
 	const char *get(size_t row_idx, size_t col_idx) const {
 		const char *row = get_row(row_idx);
@@ -374,19 +374,19 @@ public:
 		return store.get_num_nodes();
 	}
 
-	const char *get_col(off_t col_idx) const {
+	const char *get_col(size_t col_idx) const {
 		return store.get_row(col_idx);
 	}
 
-	char *get_col(off_t col_idx) {
+	char *get_col(size_t col_idx) {
 		return store.get_row(col_idx);
 	}
 
-	const char *get_cols(off_t col_start, off_t col_end) const {
+	const char *get_cols(size_t col_start, size_t col_end) const {
 		return store.get_rows(col_start, col_end);
 	}
 
-	char *get_cols(off_t col_start, off_t col_end) {
+	char *get_cols(size_t col_start, size_t col_end) {
 		return store.get_rows(col_start, col_end);
 	}
